requirements.txt
discord.py==2.3.*
humanfriendly==9.2